+++ test/e2e/hazelcast_map_custom_class_test.go
@@ -8,6 +8,7 @@ import (
 	. "time"
 
 	. "github.com/onsi/ginkgo/v2"
+	ginkgoTypes "github.com/onsi/ginkgo/v2/types"
 	. "github.com/onsi/gomega"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -38,11 +39,19 @@ var _ = Describe("Hazelcast Map Config With Custom Class Upload", Label("map"), 
 	})
 
 	AfterEach(func() {
+		GinkgoWriter.Printf("Aftereach start time is %v\n", Now().String())
+		if CurrentSpecReport().State == ginkgoTypes.SpecStateSkipped {
+			return
+		}
+		if CurrentSpecReport().State != ginkgoTypes.SpecStatePassed {
+			printDebugState()
+		}
 		DeleteAllOf(&hazelcastcomv1alpha1.Map{}, hzNamespace, labels)
 		DeleteAllOf(&hazelcastcomv1alpha1.Hazelcast{}, hzNamespace, labels)
 		DeleteAllOf(&corev1.Secret{}, hzNamespace, labels)
 		deletePVCs(hzLookupKey)
 		assertDoesNotExist(hzLookupKey, &hazelcastcomv1alpha1.Hazelcast{})
+		GinkgoWriter.Printf("Aftereach end time is %v\n", Now().String())
 	})
 
 	It("should use the MapStore implementation correctly", Label("fast"), func() {
